add atom-exec
atom-exec executes the given program with TITLE, LINK, CONTENTS, and
UPDATED set to the values of the corresponding tags in an Atom entry
fed into stdin.

Note that tags with a null byte in the contents will have a truncated
value set as a null byte is used for the string terminator.