systemd: use a calendar timer to ensure that daily updates occur
OnUnitInactiveSec is monotonic and does not increment when suspended.
Instead use a calendar timer with Persistent=true to ensure that the
feeds are updated once a day.
Unfortunately it appears that there is no fool-proof way to wait until
the network is up, so instead use an arbitrary delay in the unit file.