feed-daily: implement daily feeds
There are two types of feeds in my personal collection - those I'd like
to read on a daily basis, and those I'd prefer to batch (mostly
web graphic novels). To handle that implement a form of tagging for
feeds using a 'daily' file in the feed dir and a helper script to only
read those feeds.