atom-exec: silence spurious compile warning
The strlen is checked before the strncpy, so this was already safe.
However gcc complains; decrement the buffer size by 1 and manually null
terminate to silence to warning.
The manual null termination is not strictly required since the Feed
object is zeroed when allocated in this case, but leave in case we
someday malloc it instead.
This should exactly match the existing behaviour.