Add URL sanitization, remove remaining related FIXMEs
The main thing that needs sanitizing is input from the web - ie the feed
itself. So add URL sanitization (which makes me a bit happier about
using it in an env var for passing to a script later), but don't worry
too much about sanitizing data already in the database.